requirements.txt
pandas
openai
scikit-learn
seaborn
evaluate
wandb
tenacity
python-dotenv
google-api-core
google-genai
tiktoken
anthropic
together
transformers
torch
sentence-transformers
spacy
squarify
nltk
rouge_score
datasets
huggingface_hub
git+https://github.com/google-research/bleurt.git